PR - Merge Main Version

+++ aspnet-core/src/AtendoCloudSystem.Core/Payments/IPaymentManager.cs
@@ -17,6 +17,6 @@ namespace AtendoCloudSystem.Payments
 
         Task<Payment> UpdateAsync(Payment @payments);
 
-        Task DeleteAsync(long id);
+        Task DeleteAsync(int id);
     }
 }
